+++ src/providers/method-groups/common.ts
@@ -1,0 +1,43 @@
+import { Connector } from '../../connectors/Connector';
+import { ID, Version } from '../Types';
+
+export class JsonRpcCommonMethodGroup {
+  private conn: Connector;
+
+  constructor(conn: Connector) {
+    this.conn = conn;
+  }
+
+  /**
+   * returns peerID of libp2p node backing this API
+   */
+  public async id(): Promise<ID> {
+    const id: ID = await this.conn.request({ method: 'Filecoin.ID', params: [] });
+    return id;
+  }
+
+  /**
+   * provides information about API provider
+   */
+  public async version(): Promise<Version> {
+    const ret = await this.conn.request({ method: 'Filecoin.Version' });
+    return ret as Version;
+  }
+
+  public async logList(): Promise<string[]> {
+    const list: string[] = await this.conn.request({ method: 'Filecoin.LogList' });
+    return list;
+  }
+
+  public async logSetLevel(string1: string, string2: string): Promise<any> {
+    const result = await this.conn.request({ method: 'Filecoin.LogSetLevel', params: [string1, string2] });
+    return result;
+  }
+
+  /**
+   * trigger graceful shutdown
+   */
+  public async shutdown() {
+    await this.conn.request({ method: 'Filecoin.Shutdown' });
+  }
+}
